fix(ucmc-web): render /my/gear/cart and show add-to-cart for officers
Two bugs in the cart MVP:

1. `/my/gear/cart` was nested under `/my/gear` in the TanStack
   file-based route tree, but `my.gear.tsx` had no `<Outlet />`. So
   visiting `/my/gear/cart` actually rendered `/my/gear`'s content
   (loan list + "My cart" header button) and the cart page itself
   never reached the screen — which is what "why is there a 'my cart'
   on the my cart page" was. Split into a pathless layout
   (`my.gear.tsx` = `<Outlet />`) plus a leaf (`my.gear.index.tsx`)
   so the parent doesn't shadow the child.

2. `<AddToCartButton />` hid itself for anyone holding `gear:manage`,
   reasoning officers don't need a personal cart cluttering
   inventory. That over-extended — officers are still club members
   and may want to borrow gear themselves. Drop the permission gate;
   approved-status remains the only visibility check.
